Region coordinates updated from int to Long

+++ dbsnp-importer/src/main/java/uk/ac/ebi/eva/dbsnpimporter/models/SubSnpCoreFields.java
@@ -46,8 +46,8 @@ public class SubSnpCoreFields {
      * @param chromosomeStart Start coordinate of the variant in chromosome, null if the contig is not fully mapped to a chromosome
      * @param chromosomeEnd End coordinate of the variant in chromosome, null if the contig is not fully mapped to a chromosome
      */
-    public SubSnpCoreFields(long ssId, Long rsId, int snpOrientation, String contig, int contigStart, int contigEnd,
-                            int contigOrientation, String chromosome, Integer chromosomeStart, Integer chromosomeEnd) {
+    public SubSnpCoreFields(long ssId, Long rsId, int snpOrientation, String contig, Long contigStart, Long contigEnd,
+                            int contigOrientation, String chromosome, Long chromosomeStart, Long chromosomeEnd) {
         if (contigStart < 0 || contigEnd < 0) {
             throw new IllegalArgumentException("Contig coordinates must be non-negative numbers");
         }
@@ -87,7 +87,7 @@ public class SubSnpCoreFields {
         return contigOrientation;
     }
 
-    private Region createRegion(String sequenceName, Integer start, Integer end) {
+    private Region createRegion(String sequenceName, Long start, Long end) {
         if (sequenceName != null) {
             if (start != null) {
                 if (end != null) {
